# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Cerniq Cognitive Brain v2 API
  version: 1.0.0
  description: |
    Contract REST + convenții SSE. Ingress HTTPS via Traefik (stacks-02) pe orchestrator.
    Căi efective în `apps/api` — acest fișier trebuie păstrat sincron la schimbarea rutelor.
servers:
  - url: https://api.v2.cerniq.app
    description: Producție (Cloudflare opțional → Traefik :443 → upstream :25010)
paths:
...
  /v1/webhooks/inbound:
    post:
      summary: Webhook inbound (HMAC + idempotency)
      operationId: webhookInbound
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 8
            description: Recomandat UUID; minim 8 caractere (validare runtime)
        - name: X-Cerniq-Signature
          in: header
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Acceptat
        '400':
          description: Idempotency-Key invalid
        '401':
          description: Semnătură invalidă
        '503':
          description: Secret webhook neconfigurat
        '409':
          description: Idempotency replay
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...